+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+### Changed
+
+- Admin: Make the order editor keep the suppliers of non-product order lines intact. 
+
+### Fixed:
+
+- Admin: Fix the edit button on the order editor.
+
+
 ## [2.0.5] - 2020-08-16
 
 ### Added